Mejorando visualizacion y direcciones
