+++ src/app/miembros/colaborar/colaborar.module.css
@@ -34,6 +34,7 @@
 
 .li {
   background: #f9f9f9;
+  color:#000000;
   padding: 15px;
   margin-bottom: 10px;
   border-radius: 8px;
