+++ src/app/miembros/tareas/tareas.module.css
@@ -34,6 +34,7 @@
 
 .taskItem {
   background: #f9f9f9;
+  color:#000000;
   padding: 15px;
   margin-bottom: 10px;
   border-radius: 8px;
